ems-backend/app/main.py
from fastapi import FastAPI
from starlette.middleware.cors import CORSMiddleware

app = FastAPI(title="Employee Management System")

# Set all CORS enabled origins
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

from app.api.api import api_router
from app.core.config import settings

app.include_router(api_router, prefix=settings.API_V1_STR)

@app.get("/")
def read_root():
    return {"message": "Welcome to the Employee Management System API"}